+++ common/db/src/main/java/is/codion/common/db/pool/DefaultConnectionPoolCounter.java
@@ -44,7 +44,8 @@ final class DefaultConnectionPoolCounter {
 	private final LinkedList<DefaultConnectionPoolState> snapshotStatistics = new LinkedList<>();
 	private volatile boolean collectSnapshotStatistics = false;
 	private volatile boolean collectCheckOutTimes = false;
-	private final TaskScheduler snapshotStatisticsCollector = TaskScheduler.builder(new StatisticsCollector())
+	private final TaskScheduler snapshotStatisticsCollector = TaskScheduler.builder()
+					.task(new StatisticsCollector())
 					.interval(SNAPSHOT_COLLECTION_INTERVAL_MS, TimeUnit.MILLISECONDS)
 					.build();