+++ tools/loadtest/ui/src/main/java/is/codion/tools/loadtest/ui/LoadTestPanel.java
@@ -537,8 +537,8 @@ public final class LoadTestPanel<T> extends JPanel {
 										.columns(8)
 										.editable(false)
 										.horizontalAlignment(SwingConstants.CENTER)
-										.onBuild(memoryUsageField -> TaskScheduler.builder(() -> SwingUtilities.invokeLater(() ->
-																		memoryUsageField.setText(memoryUsage())))
+										.onBuild(memoryUsageField -> TaskScheduler.builder()
+														.task(() -> SwingUtilities.invokeLater(() -> memoryUsageField.setText(memoryUsage())))
 														.interval(DEFAULT_MEMORY_USAGE_UPDATE_INTERVAL_MS, TimeUnit.MILLISECONDS)
 														.start())
 										.build())