+++ tools/monitor/model/src/main/java/is/codion/tools/monitor/model/ClientUserMonitor.java
@@ -78,7 +78,8 @@ public final class ClientUserMonitor {
 						.value(getIdleConnectionTimeout())
 						.consumer(this::setIdleConnectionTimeout)
 						.build();
-		this.updateScheduler = TaskScheduler.builder(this::refreshUserHistoryTableModel)
+		this.updateScheduler = TaskScheduler.builder()
+						.task(this::refreshUserHistoryTableModel)
 						.interval(updateRate, TimeUnit.SECONDS)
 						.start();
 	}