+++ tools/monitor/model/src/main/java/is/codion/tools/monitor/model/ServerMonitor.java
@@ -171,7 +171,8 @@ public final class ServerMonitor {
 		this.systemLoadCollection.addSeries(processLoadSeries);
 		this.databaseMonitor = new DatabaseMonitor(server, updateRate);
 		this.clientMonitor = new ClientUserMonitor(server, updateRate);
-		this.updateScheduler = TaskScheduler.builder(this::updateStatistics)
+		this.updateScheduler = TaskScheduler.builder()
+						.task(this::updateStatistics)
 						.interval(updateRate, TimeUnit.SECONDS)
 						.start();
 		refreshDomainList();